use dbversion=71 for updating ConfiguredProvider
in addition to merged pr #2123,
dbversion=70 was also used by pr #2125.

pr #2125 is not yet merged, however beta-versions were shipped with
that pr, so, dbversion=70 is already in use on some systems and on these
systems ConfiguredProvider won't be set.

the other way round, that ConfiguredProvider is set twice on some systems
affects only dev-installations and should not harm.

once this pr is merged, #2125 needs a rebase and will use dbversion=72 for
its changes to add the new column also n the dev-installations :)
(we need to check that the column to add does not exist, however)