Remove obsolete ChatId::is_unset check
The get_chat_id_by_grpid() call is already resultified.  The database
can only contain consistent data since on creation it inserts rows for
the special chats which do not have a grpid set so can never match.
Thus we can safely remove another use of the ChatId::is_unset call
which should be removed to clean up the type.