do not re-use bind parameters
sqlx does not bind the first `?` implicitly to `?1`,
therefore the sql-statement fails.

we could use `?1` several times, however,
i did not find good documentation about that,
and even if that works now, that could fail with other db
or even with sqlite in newer versions.

therefore, we just bind every parameter explicitly.