Increase minimal origin for contact list
Currently contact list contains contacts from all groups, which makes
it cluttered with every contact, including bots, that has ever been in
the same group as you.

For search, minimal origin remains the same as before, so it is possible
to find anyone you want by searching, for example if you want to create
a new group but don't want to start 1:1 chat with every member first.