Don't send messages in the self chat to self if bcc_self is off
When writing a UI regression test for
https://github.com/deltachat/deltachat-android/issues/1704, I had the
problem that all sent messages stay in state OutPending, and the UI
shows a rotating dotted circle in the lower right corner.

However, after clicking a button, the testing framework (Espresso)
always waits until the UI thread is IDLEing (which means that nothing on
the screen moves anymore) before continuing. The rotating circle means
that it waits forever.

We don't have to change this just because it would be useful for the UI
test, but I think that it makes sense anyway that messages in "Saved
Messages" aren't the only messages that are stored on the server when
`bcc_self` is off.